% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csp}
\alias{csp}
\title{Correlates of State Policy Dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4385745 rows and 13 columns.}
\usage{
csp
}
\description{
This package contains the Correlates of State Policy dataset, which includes
more than seven-hundred variables, with observations across the U.S. 50
states and time (1900 - 2016).
}
\details{
For more information, go to the
\href{http://ippsr.msu.edu/public-policy/correlates-state-policy}{official website}.
}
\examples{
\dontrun{
library(csp)
data(csp, package = "csp")
dim(csp)
names(csp)
csp[1:2, ]}
}
\keyword{datasets}

